refactor: 스크롤 관련 비즈니스 코드 별도의 커스텀 훅으로 분리 #273
